refactor: add FROM and TO for make consults

+++ server/SNMP/main.sh
@@ -3,12 +3,18 @@ spinner=("Getting consults SNMP... -" "Getting consults SNMP... \\" "Getting con
 current_date=$(date +%Y-%m-%d)
 communities=()
 
-CURRENTPATH="$(pwd)/Interface-Change-Monitor/server/SNMP"
+FROM=$1
+TO=$2
+PART=$3
+USERNAME=$(echo $USER)
+CURRENTPATH="$(echo $HOME)/Interface-Change-Monitor/server/SNMP"
 PING="$CURRENTPATH/ping.sh"
 SNMP="$CURRENTPATH/consult.sh"
-INPUT="$CURRENTPATH/devices.csv"
-OUTPUT="$CURRENTPATH/data/SNMP_$current_date"
+INPUT="$CURRENTPATH/devices.txt"
+OUTPUT="$CURRENTPATH/data/SNMP_"$current_date"_part_$PART"
 
+##########################################################
+# Load Spinner
 spinner() {
     while :; do
         for i in "${spinner[@]}"; do
@@ -17,22 +23,39 @@ spinner() {
         done
     done
 }
+#########################################################
+# Check if the "FROM" and "TO" are correct values
+if [ $TO -lt $FROM ]; then
+    echo "Bad values of 'from' and 'to' input."
+    exit 1
+fi
 
-echo "Time started: $(date)"
-
+# Check of the data file (devices)
 if [ ! -f $INPUT ]; then
     echo "Data file not found."
     exit 1
 fi
 
+# Delete other version of consults SNMP with the same date
 if [ -f $OUTPUT ]; then
     rm $OUTPUT
 fi
 
-spinner &
-spinner_pid=$!
+# If the script is run manually, the load spinner will come up.
+if [ -n "$USERNAME" ]; then 
+    spinner &
+    spinner_pid=$!
+fi
 
-# BEGIN
+# Check if this is the first consult
+if [ $FROM != 0 ]; then
+    FROM=$((FROM + 1))
+fi
+#############################################################
+# Start time registration
+echo "Time started: $(date)"
+
+# Programm Begin
 while IFS=, read -r col1 col2
 do
     ip="$col1"
@@ -41,9 +64,9 @@ do
     communities+=("$community")
 done < "$INPUT"
 
-total_communities=${#communities[@]}
+consult_realized=0
 
-for (( i = 0; i < $total_communities; i+=2)); do
+for (( i = $FROM; consult_realized <= $TO; i+=2)); do
     j=$(( i + 1 ))
     ip=${communities[$i]}
     community=${communities[$j]}
@@ -51,9 +74,15 @@ for (( i = 0; i < $total_communities; i+=2)); do
     if [ -n "$response" ]; then
         bash $SNMP $OUTPUT $community $ip $current_date
     fi
+    consult_realized=$((consult_realized + 1))
 done
-# END
+# Programm End
 
-kill $spinner_pid
-wait $spinner_pid
-echo "Time finished: $(date)"
+# If the script is run manually, the load spinner will come down.
+if [ -n "$USERNAME" ]; then 
+    kill $spinner_pid
+    wait $spinner_pid
+fi
+
+# End time registration
+echo -e "\nTime finished: $(date)"
